+++ src/main/java/org/truffleruby/options/Options.java
@@ -132,6 +132,8 @@ public class Options {
     public final boolean CEXTS_LOG_WARNINGS;
     /** --argv-globals=false */
     public final boolean ARGV_GLOBALS;
+    /** --chomp-loop=false */
+    public final boolean CHOMP_LOOP;
     /** --gets-loop=false */
     public final boolean GETS_LOOP;
     /** --print-loop=false */
@@ -327,6 +329,7 @@ public class Options {
         CEXTS_LOG_LOAD = options.get(OptionsCatalog.CEXTS_LOG_LOAD_KEY);
         CEXTS_LOG_WARNINGS = options.get(OptionsCatalog.CEXTS_LOG_WARNINGS_KEY);
         ARGV_GLOBALS = options.get(OptionsCatalog.ARGV_GLOBALS_KEY);
+        CHOMP_LOOP = options.get(OptionsCatalog.CHOMP_LOOP_KEY);
         GETS_LOOP = options.get(OptionsCatalog.GETS_LOOP_KEY);
         PRINT_LOOP = options.get(OptionsCatalog.PRINT_LOOP_KEY);
         SPLIT_LOOP = options.get(OptionsCatalog.SPLIT_LOOP_KEY);
@@ -510,6 +513,8 @@ public class Options {
                 return CEXTS_LOG_WARNINGS;
             case "ruby.argv-globals":
                 return ARGV_GLOBALS;
+            case "ruby.chomp-loop":
+                return CHOMP_LOOP;
             case "ruby.gets-loop":
                 return GETS_LOOP;
             case "ruby.print-loop":