+++ src/main/ruby/truffleruby/core/truffle/kernel_operations.rb
@@ -43,7 +43,7 @@ module Truffle
     define_read_only_global(:$*, -> { ARGV })
 
     define_read_only_global(:$-a, -> { Truffle::Boot.get_option 'split-loop' })
-    define_read_only_global(:$-l, -> { nil })
+    define_read_only_global(:$-l, -> { Truffle::Boot.get_option 'chomp-loop' })
     define_read_only_global(:$-p, -> { Truffle::Boot.get_option 'print-loop' })
 
     define_hooked_variable(
@@ -57,6 +57,12 @@ module Truffle
       })
 
     $/ = "\n".freeze
+
+    Truffle::Boot.delay do
+      if Truffle::Boot.get_option 'chomp-loop'
+        $\ = $/
+      end
+    end
 
     alias $-0 $/